Step 5 - Device Injestion for Gateway API - renamed and ccleaned unsed libariy refernce to avoid deployment config due to surface pro 13 machine

+++ src/NewDoor.Presentation/NewDoor.DeviceSimulator/Models/TelemetryPayload.cs
@@ -8,7 +8,7 @@ public class DeviceTelemetryPayload
     public string Status { get; set; } = string.Empty;
     public int BatteryLevel { get; set; }
     public double? Temperature { get; set; }
-    public int? SmokeLevel { get; set; }
+    public double? SmokeLevel { get; set; }
     public string? SignalStrength { get; set; }
 }
